# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Labour Tracker",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/subscription/subscribe/{labour_id}": {
            "post": {
                "tags": [
                    "Subscription"
                ],
                "summary": "Subscribe To",
                "operationId": "subscribe_to_api_v1_subscription_subscribe__labour_id__post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "labour_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Labour Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubscribeToRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubscriptionResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/subscription/unsubscribe": {
            "post": {
                "tags": [
                    "Subscription"
                ],
                "summary": "Unsubscribe From",
                "operationId": "unsubscribe_from_api_v1_subscription_unsubscribe_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UnsubscribeFromRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubscriptionResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/subscription/subscriptions": {
            "get": {
                "tags": [
                    "Subscription"
                ],
                "summary": "Get Subscriptions",
                "operationId": "get_subscriptions_api_v1_subscription_subscriptions_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubscriptionsResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/subscription/labour_subscriptions/{labour_id}": {
            "get": {
                "tags": [
                    "Subscription"
                ],
                "summary": "Get Labour Subscriptions",
                "operationId": "get_labour_subscriptions_api_v1_subscription_labour_subscriptions__labour_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "labour_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Labour Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LabourSubscriptionsResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ExceptionSchema": {
                "properties": {
                    "description": {
                        "type": "string",
                        "title": "Description"
                    }
                },
                "type": "object",
                "required": [
                    "description"
                ],
                "title": "ExceptionSchema"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "LabourSubscriptionsResponse": {
                "properties": {
                    "subscriptions": {
                        "items": {
                            "$ref": "#/components/schemas/SubscriptionDTO"
                        },
                        "type": "array",
                        "title": "Subscriptions"
                    },
                    "subscribers": {
                        "items": {
                            "$ref": "#/components/schemas/SubscriberDTO"
                        },
                        "type": "array",
                        "title": "Subscribers"
                    }
                },
                "type": "object",
                "required": [
                    "subscriptions",
                    "subscribers"
                ],
                "title": "LabourSubscriptionsResponse"
            },
...
            "SubscribeToRequest": {
                "properties": {
                    "token": {
                        "type": "string",
                        "title": "Token"
                    }
                },
                "type": "object",
                "required": [
                    "token"
                ],
                "title": "SubscribeToRequest"
            },
            "SubscriberDTO": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "first_name": {
                        "type": "string",
                        "title": "First Name"
                    },
                    "last_name": {
                        "type": "string",
                        "title": "Last Name"
                    },
                    "phone_number": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Phone Number"
                    },
                    "email": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Email"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "first_name",
                    "last_name",
                    "phone_number",
                    "email"
                ],
                "title": "SubscriberDTO"
            },
...
            "SubscriptionDTO": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "labour_id": {
                        "type": "string",
                        "title": "Labour Id"
                    },
                    "birthing_person_id": {
                        "type": "string",
                        "title": "Birthing Person Id"
                    },
                    "subscriber_id": {
                        "type": "string",
                        "title": "Subscriber Id"
                    },
                    "role": {
                        "type": "string",
                        "title": "Role"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "contact_methods": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Contact Methods"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "labour_id",
                    "birthing_person_id",
                    "subscriber_id",
                    "role",
                    "status",
                    "contact_methods"
                ],
                "title": "SubscriptionDTO"
            },
            "SubscriptionResponse": {
                "properties": {
                    "subscription": {
                        "$ref": "#/components/schemas/SubscriptionDTO"
                    }
                },
                "type": "object",
                "required": [
                    "subscription"
                ],
                "title": "SubscriptionResponse"
            },
            "SubscriptionsResponse": {
                "properties": {
                    "subscriptions": {
                        "items": {
                            "$ref": "#/components/schemas/SubscriptionDTO"
                        },
                        "type": "array",
                        "title": "Subscriptions"
                    }
                },
                "type": "object",
                "required": [
                    "subscriptions"
                ],
                "title": "SubscriptionsResponse"
            },
...
            "UnsubscribeFromRequest": {
                "properties": {
                    "labour_id": {
                        "type": "string",
                        "title": "Labour Id"
                    }
                },
                "type": "object",
                "required": [
                    "labour_id"
                ],
                "title": "UnsubscribeFromRequest"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}